prac4

+++ bus.cpp
@@ -5,5 +5,5 @@ using namespace std;
 Bus::Bus(int ID): Vehicle(ID){};
 
 int Bus::getParkingDuration(){
-  return 0.75*this->Vehicle::getParkingDuration();
+  return int(0.75*this->Vehicle::getParkingDuration());
 }
