+++ car.cpp
@@ -5,5 +5,5 @@ using namespace std;
 Car::Car(int ID): Vehicle(ID){};
 
 int Car::getParkingDuration(){
-  return 0.9*this->Vehicle::getParkingDuration();
+  return int(0.9*this->Vehicle::getParkingDuration());
 }
