+++ motorbike.cpp
@@ -5,5 +5,5 @@ using namespace std;
 Motorbike::Motorbike(int ID): Vehicle(ID){};
 
 int Motorbike::getParkingDuration(){
-  return 0.85 * this->Vehicle::getParkingDuration();
+  return int(0.85 * this->Vehicle::getParkingDuration());
 }
